+++ Apps/Reinstall AppX Packages.ps1
@@ -4,16 +4,18 @@
 .SYNOPSIS
   Re-registers all packaged apps by Microsoft Corporation for the current user.
 .DESCRIPTION
-  Looks for all mainline packaged apps whose publishers are "cw5n1h2txyewy" (Microsoft apps) or
-  "8wekyb3d8bbwe" (Windows components). Re-registers them for the current user.
+  This script queries all mainline packaged apps whose publishers are "cw5n1h2txyewy" (Microsoft
+  apps) or "8wekyb3d8bbwe" (Windows components). Re-registers them for the current user.
 .NOTES
-  DEPRECATED
+  DEPRECATED. I strongly recommend not to use it.
 
   This script depends on Get-AppxPackage and its Appx module. As such, it won't work in PowerShell
-  6 and later.
+  6 and later, and returns an error (System.PlatformNotSupportedException).
 .LINK
   None
 #>
+
+Import-Module -Name Appx -ErrorAction "Stop"
 
 $MicrosoftPublisherIDs = @(
     "cw5n1h2txyewy", # CN=Microsoft Corporation, O=Microsoft Corporation, L=Redmond, S=Washington, C=US